Start Weston in its dedicated user $HOME directory
If systemd runs Weston as a system-wide service (which is
the case here), its working directory will be "/".

It is a problem for privilegied clients such as the screen
recorder and screenshooter, which will try to dump their
output files there, and fail. Make them create these in
"/run/%user" instead.

Bug-Tizen: TC-1359

Change-Id: Id6322f38f8dcbec4efaa68906a2fef27815c93d1
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>